# docs/bruno/NexusPay/apispec/api.yaml
openapi: 3.0.0
info:
  title: api
  version: 1.0.0
paths:
  /auth/login:
    post:
      summary: Login user
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_authentication_login_user_yml
      description: ''
      tags:
        - Authentication
      responses:
        '200':
          description: OK. Refresh token is set as an HTTP-only cookie.
          headers:
            Set-Cookie:
              description: >-
                refresh_token=<token>; Path=/; HttpOnly; SameSite=Lax.
                Expires after 7 days.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  full_name:
                    type: string
                  jwt_token:
                    type: string
              examples:
                200_Response:
                  value:
                    email: john.doe@example.com
                    full_name: John Doe
                    jwt_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  summary: 200 Response
                  description: Login successful
                example:
                  value:
                    email: youssef@gmail.com
                    full_name: Youssef Hany
                    jwt_token: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NzgxMTM5MTAsImlhdCI6MTc3ODExMzAxMCwic3ViIjoiNzYzNzQ2ZjgtY2EyOC00OWE2LTkyZjUtMjhlN2YwOWVhMTFjIn0.hF2HnlJMflAosc4TN9lOwMvXyYZW0maM52mX_lqwj6c
                  summary: example
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: invalid credentials
                message: The provided credentials are incorrect
                status_code: 401
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: user not found
                message: No user found with this email
                status_code: 404
      requestBody:
        $ref: '#/components/requestBodies/login_user'
  /users/logout:
    post:
      summary: Logout user
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_authentication_logout_user_yml
      description: ''
      tags:
        - Authentication
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401
      security:
        - logout_user: []
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_authentication_refresh_access_token_yml
      description: ''
      tags:
        - Authentication
      responses:
        '200':
          description: OK. Rotated refresh token is set as a new HTTP-only cookie.
          headers:
            Set-Cookie:
              description: >-
                refresh_token=<new_token>; Path=/; HttpOnly; SameSite=Lax.
                Expires after 7 days.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt_token:
                    type: string
              examples:
                200_Response:
                  value:
                    jwt_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  summary: 200 Response
                  description: Token refreshed successfully
                example:
                  value:
                    jwt_token: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NzgxMTM5MTAsImlhdCI6MTc3ODExMzAxMCwic3ViIjoiNzYzNzQ2ZjgtY2EyOC00OWE2LTkyZjUtMjhlN2YwOWVhMTFjIn0.hF2HnlJMflAosc4TN9lOwMvXyYZW0maM52mX_lqwj6c
                  summary: example
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: User not found
                message: Invalid or expired refresh token
                status_code: 404
  /auth/register:
    post:
      summary: Register a new user
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_authentication_register_a_new_user_yml
      description: ''
      tags:
        - Authentication
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  full_name:
                    type: string
                  jwt_token:
                    type: string
                  refresh_token:
                    type: string
              examples:
                201_Response:
                  value:
                    email: john.doe@example.com
                    full_name: John Doe
                    jwt_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    refresh_token: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
                  summary: 201 Response
                  description: User successfully registered
                example:
                  value:
                    email: newuser@example.com
                    full_name: New User
                    jwt_token: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NzgxMTM5MTAsImlhdCI6MTc3ODExMzAxMCwic3ViIjoiNzYzNzQ2ZjgtY2EyOC00OWE2LTkyZjUtMjhlN2YwOWVhMTFjIn0.hF2HnlJMflAosc4TN9lOwMvXyYZW0maM52mX_lqwj6c
                    refresh_token: >-
                      79bbfb367e6bdb646cbd63a4d67058ac4de02ad35612c54bdc56144c97e8f167
                  summary: example
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: Invalid input
                message: Email is required and must be valid
                status_code: 400
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: User already exists
                message: A user with this email already exists
                status_code: 409
      requestBody:
        $ref: '#/components/requestBodies/register_a_new_user'
      servers:
        - url: http://localhost:3000
  /users/test:
    get:
      summary: Test authentication
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_authentication_test_authentication_yml
      description: ''
      tags:
        - Authentication
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  exp:
                    type: integer
                  iat:
                    type: integer
              examples:
                200_Response:
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    exp: 1705317600
                    iat: 1704712800
                  summary: 200 Response
                  description: Authentication successful, returns token claims
                example:
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    exp: 1705317600
                    iat: 1704712800
                  summary: example
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401
      security:
        - test_authentication: []
...
  /transfers/:
    post:
      summary: Create transfer
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_transfers_create_transfer_yml
      description: ''
      tags:
        - Transfers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  from_wallet_id:
                    type: string
                  to_wallet_id:
                    type: string
                  amount_in_piastres:
                    type: integer
                  status:
                    type: string
                  note:
                    type: string
                  created_at:
                    type: string
              examples:
                200_Response:
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174001
                    from_wallet_id: 123e4567-e89b-12d3-a456-426614174000
                    to_wallet_id: 123e4567-e89b-12d3-a456-426614174002
                    amount_in_piastres: 5000
                    status: completed
                    note: Dinner split
                    created_at: '2024-01-15T10:30:00Z'
                  summary: 200 Response
                  description: Transfer created successfully
                200_Response_Scheduled:
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174001
                    from_wallet_id: 123e4567-e89b-12d3-a456-426614174000
                    to_wallet_id: 123e4567-e89b-12d3-a456-426614174002
                    amount_in_piastres: 5000
                    status: pending
                    note: Rent payment
                    created_at: '2024-01-15T10:30:00Z'
                  summary: 200 Response (Scheduled)
                  description: Scheduled transfer created successfully
                example:
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174001
                    from_wallet_id: 023e4567-e89b-12d3-a456-426614174000
                    to_wallet_id: 123e4567-e89b-12d3-a456-426614174002
                    amount_in_piastres: 5000
                    status: completed
                    note: Dinner split
                    created_at: '2024-01-15T10:30:00Z'
                  summary: example
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              examples:
                400_Response:
                  value:
                    error: Bad request
                    message: Amount must be at least 1000 piastres
                    status_code: 400
                  summary: 400 Response
                  description: Bad request - invalid input
                400_Response_2:
                  value:
                    error: Bad request
                    message: Can not transfer to self
                    status_code: 400
                  summary: 400 Response
                  description: Bad request - cannot transfer to self
                400_Response_3:
                  value:
                    error: Bad request
                    message: Insufficient funds
                    status_code: 400
                  summary: 400 Response
                  description: Bad request - insufficient funds
                400_Response_4:
                  value:
                    error: Bad request
                    message: Scheduled time must be in the future
                    status_code: 400
                  summary: 400 Response
                  description: Bad request - scheduled time must be in future
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: Wallet not found
                message: Wallet not found
                status_code: 404
      requestBody:
        $ref: '#/components/requestBodies/create_transfer'
      security:
        - create_transfer: []
    get:
      summary: Get transfers
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_transfers_get_transfers_yml
      description: ''
      tags:
        - Transfers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transfers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        from_wallet_id:
                          type: string
                        from_user:
                          type: object
                          properties:
                            id:
                              type: string
                            full_name:
                              type: string
                        to_wallet_id:
                          type: string
                        to_user:
                          type: object
                          properties:
                            id:
                              type: string
                            full_name:
                              type: string
                        amount_in_piastres:
                          type: integer
                        direction:
                          type: string
                        status:
                          type: string
                        note:
                          type: string
                        created_at:
                          type: string
              examples:
                200_Response:
                  value:
                    transfers:
                      - id: 123e4567-e89b-12d3-a456-426614174001
                        from_wallet_id: 123e4567-e89b-12d3-a456-426614174000
                        from_user:
                          id: 123e4567-e89b-12d3-a456-426614174000
                          full_name: Alice Smith
                        to_wallet_id: 123e4567-e89b-12d3-a456-426614174002
                        to_user:
                          id: 123e4567-e89b-12d3-a456-426614174003
                          full_name: John Doe
                        amount_in_piastres: 5000
                        direction: debit
                        status: completed
                        note: Dinner split
                        created_at: '2024-01-15T10:30:00Z'
                      - id: 123e4567-e89b-12d3-a456-426614174004
                        from_wallet_id: 123e4567-e89b-12d3-a456-426614174000
                        from_user:
                          id: 123e4567-e89b-12d3-a456-426614174000
                          full_name: Alice Smith
                        to_wallet_id: 123e4567-e89b-12d3-a456-426614174005
                        to_user:
                          id: 123e4567-e89b-12d3-a456-426614174006
                          full_name: Jane Smith
                        amount_in_piastres: 3000
                        direction: credit
                        status: completed
                        note: Lunch split
                        created_at: '2024-01-16T14:20:00Z'
                  summary: 200 Response
                  description: Transfers retrieved successfully
                example:
                  value:
                    transfers:
                      - id: 123e4567-e89b-12d3-a456-426614174001
                        from_wallet_id: 023e4567-e89b-12d3-a456-426614174000
                        from_user:
                          id: 023e4567-e89b-12d3-a456-426614174000
                          full_name: Alice Smith
                        to_wallet_id: 223e4567-e89b-12d3-a456-426614174002
                        to_user:
                          id: 323e4567-e89b-12d3-a456-426614174003
                          full_name: John Doe
                        amount_in_piastres: 5000
                        direction: debit
                        status: completed
                        note: Dinner split
                        created_at: '2024-01-15T10:30:00Z'
                      - id: 423e4567-e89b-12d3-a456-426614174004
                        from_wallet_id: 123e4567-e89b-12d3-a456-426614174000
                        from_user:
                          id: 123e4567-e89b-12d3-a456-426614174000
                          full_name: Alice Smith
                        to_wallet_id: 523e4567-e89b-12d3-a456-426614174005
                        to_user:
                          id: 623e4567-e89b-12d3-a456-426614174006
                          full_name: Jane Smith
                        amount_in_piastres: 3000
                        direction: credit
                        status: completed
                        note: Lunch split
                        created_at: '2024-01-16T14:20:00Z'
                  summary: example
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401
      security:
        - get_transfers: []
...
  /transfers/scheduled/:
    get:
      summary: Get scheduled transfers
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_transfers_scheduled_get_scheduled_transfers_yml
      description: ''
      tags:
        - Transfers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  scheduled_transfers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        transfer_id:
                          type: string
                        scheduled_at:
                          type: string
                        executed_at:
                          type: string
                          nullable: true
                        deleted_at:
                          type: string
                          nullable: true
                        created_at:
                          type: string
              examples:
                200_Response:
                  value:
                    scheduled_transfers:
                      - id: 123e4567-e89b-12d3-a456-426614174001
                        transfer_id: 123e4567-e89b-12d3-a456-426614174002
                        scheduled_at: '2024-01-20T10:00:00Z'
                        executed_at: null
                        deleted_at: null
                        created_at: '2024-01-15T10:30:00Z'
                  summary: 200 Response
                  description: Scheduled transfers retrieved successfully
                example:
                  value:
                    scheduled_transfers:
                      - id: 123e4567-e89b-12d3-a456-426614174001
                        transfer_id: 223e4567-e89b-12d3-a456-426614174002
                        scheduled_at: '2024-01-20T10:00:00Z'
                        executed_at: null
                        deleted_at: null
                        created_at: '2024-01-15T10:30:00Z'
                  summary: example
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401
      security:
        - get_scheduled_transfers: []
...
  /users?name=John:
    get:
      summary: Search users
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_users_search_users_yml
      description: ''
      tags:
        - Users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        full_name:
                          type: string
              example:
                users:
                  - id: 123e4567-e89b-12d3-a456-426614174001
                    full_name: John Doe
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: unauthorized
                message: Missing or invalid authorization token
                status_code: 401
      parameters:
        - name: name
          in: query
          description: ''
          required: true
          schema:
            type: string
          example: John
      security:
        - search_users: []
...
  /webhook/stripe:
    post:
      summary: Stripe webhook
      operationId: >-
        home_joe-arch_projects_go_nexus_docs_bruno_nexuspay_collections_nexusapii_webhooks_stripe_webhook_yml
      description: ''
      tags:
        - Webhooks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
              examples:
                200_Response:
                  value:
                    received: true
                  summary: 200 Response
                  description: Webhook handled successfully
                example:
                  value:
                    received: true
                  summary: example
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  status_code:
                    type: integer
              example:
                error: invalid signature
                message: Webhook signature verification failed
                status_code: 400
      parameters:
        - name: Stripe-Signature
          in: header
          description: ''
          required: true
          schema:
            type: string
          example: whsec_test_signature
      requestBody:
        $ref: '#/components/requestBodies/stripe_webhook'
servers:
  - url: http://localhost:3000
    description: Base Server
  - url: http://localhost:3000
    description: Development server
  - url: api.nexuspay.com
    description: Production server
components:
  schemas:
    login_user:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: youssef@gmail.com
        password: passwrd
    register_a_new_user:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        full_name:
          type: string
      example:
        email: john@gmail.com
        password: password
        full_name: John Doe
    create_transfer:
      type: object
      properties:
        to_wallet_id:
          type: string
        amount_in_piastres:
          type: integer
        note:
          type: string
      example:
        to_wallet_id: 123e4567-e89b-12d3-a456-426614174002
        amount_in_piastres: 5000
        note: Dinner split
...
  requestBodies:
    login_user:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/login_user'
      description: ''
      required: true
    register_a_new_user:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/register_a_new_user'
      description: ''
      required: true
    create_transfer:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/create_transfer'
      description: ''
      required: true
...
    stripe_webhook:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/stripe_webhook'
      description: ''
      required: true
  securitySchemes:
    logout_user:
      type: http
      scheme: bearer
    test_authentication:
      type: http
      scheme: bearer
    create_transfer:
      type: http
      scheme: bearer
    get_transfers:
      type: http
      scheme: bearer
    get_transfer_by_id:
      type: http
      scheme: bearer
    get_scheduled_transfers:
      type: http
      scheme: bearer
    cancel_scheduled_transfer:
      type: http
      scheme: bearer
    search_users:
      type: http
      scheme: bearer
    get_payments:
      type: http
      scheme: bearer
    get_wallet:
      type: http
      scheme: bearer
    top_up_wallet:
      type: http
      scheme: bearer